+++ NEWS.md
@@ -14,6 +14,7 @@ Minimum LaBB-CAT version *20221111.1200*
 - *getMedia* now downloads the media and returns the file name, rather than returning the URL.
 - *getMatches* accepts a string as a pattern, and assumes it's a space-delimited orthography layer search.
 - *getMatches* now has *anchor.confidence.min* parameter, to allow retrieval of start/end times that have low confidence.
+- *getMatchLabels* and *getMatchAlignments* now return `NA` rather than empty strings.
 
 # nzilbb.labbcat 1.1-1